db: move host benchmarks
